# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KARBONİZGİ",
        "contact": {}
    },
    "host": "https://api.karbonizgi.tr",
    "paths": {
        "/add-suggest": {
            "post": {
                "description": "Aksiyon Ekleme Kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Action"
                ],
                "summary": "Kullanıcı aksiyon ekleme",
                "parameters": [
                    {
                        "description": "status değeri için sadece 'planned,in_progress,completed,cancelled' değerlerden biri gönderilmelidir. ",
                        "name": "Suggest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Suggest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
        "/delete-suggest": {
            "post": {
                "description": "Aksiyon Silme Kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Action"
                ],
                "summary": "Kullanıcı aksiyon silme",
                "parameters": [
                    {
                        "description": "action id gönderilmesi yeterlidir.",
                        "name": "Suggest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DelSuggest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
        "/score-rank": {
            "get": {
                "description": "Score Ranking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Score"
                ],
                "summary": "Kullanıcı skor sıralaması",
                "responses": {
                    "200": {
                        "description": "total_score_number toplam skor tablosundaki sayıyı gösterir. score_number kullanıcının sıralamasını score_ranking ise ilk 4 sıralamayı gösterir. rank sıralamadır.",
                        "schema": {
                            "$ref": "#/definitions/controllers.RankResp"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
        "/update-suggest": {
            "post": {
                "description": "Aksiyon Güncelleme Kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Action"
                ],
                "summary": "Kullanıcı aksiyon güncelleme",
                "parameters": [
                    {
                        "description": "status değeri için sadece 'planned,in_progress,completed,cancelled' değerlerden biri gönderilmelidir. ",
                        "name": "Suggest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UptSuggest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.DelSuggest": {
            "type": "object",
            "properties": {
                "action_id": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.RankResp": {
            "type": "object",
            "properties": {
                "score_number": {
                    "type": "integer"
                },
                "score_ranking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.UserRank"
                    }
                },
                "total_score_number": {
                    "type": "integer"
                }
            }
        },
        "controllers.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.Suggest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UptSuggest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "action_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UserRank": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                }
            }
        },
...
    }
}